+++ lib/markdown/index.ts
@@ -10,4 +10,6 @@ export type {
 export * from './markdown-parser';
 export * from './markdown-serializer';
 export * from './todo-list-markdown-it-plugin';
+export * from './inline-node-parser';
+
 export type { Token, StateCore };